Made the details section responsive
Removed the faux table layout made using nbsp and replaced it with a
proper HTML table that will render correctly on all screen sizes.
Will also help with accessibility, screen readers, and keyboard
navigation as the table will now be read correctly as per HTML
semantics.

Made the font sizes smaller for the table items on the mobile, and
retained the bigger size on the desktop.